requirements.txt
python-telegram-bot==21.0
aiohttp==3.9.0
beautifulsoup4==4.12.0
lxml==5.1.0
